chore: resolve PR convos
