+++ src/plugins/tokenPlugin/types/tokenMember.ts
@@ -18,8 +18,4 @@ export interface ITokenMember extends Omit<IMember, 'metrics'> {
      * Metrics for the token member.
      */
     metrics: ITokenMemberMetrics;
-    /**
-     * Current user delegate.
-     */
-    currentDelegate: string | null;
 }
